+++ source/common/quic/quic_client_transport_socket_factory.cc
@@ -10,4 +10,57 @@
 
 #include "quiche/quic/core/crypto/quic_client_session_cache.h"
 
-namespace Envoy {} // namespace Envoy
+namespace Envoy {
+namespace Quic {
+
+Network::UpstreamTransportSocketFactoryPtr
+QuicClientTransportSocketConfigFactory::createTransportSocketFactory(
+    const Protobuf::Message& config,
+    Server::Configuration::TransportSocketFactoryContext& context) {
+  auto quic_transport = MessageUtil::downcastAndValidate<
+      const envoy::extensions::transport_sockets::quic::v3::QuicUpstreamTransport&>(
+      config, context.messageValidationVisitor());
+  auto client_config = std::make_unique<Extensions::TransportSockets::Tls::ClientContextConfigImpl>(
+      quic_transport.upstream_tls_context(), context);
+  auto factory =
+      std::make_unique<QuicClientTransportSocketFactory>(std::move(client_config), context);
+  factory->initialize();
+  return factory;
+}
+
+QuicClientTransportSocketFactory::QuicClientTransportSocketFactory(
+    Ssl::ClientContextConfigPtr config,
+    Server::Configuration::TransportSocketFactoryContext& factory_context)
+    : QuicTransportSocketFactoryBase(factory_context.statsScope(), "client"),
+      fallback_factory_(std::make_unique<Extensions::TransportSockets::Tls::ClientSslSocketFactory>(
+          std::move(config), factory_context.sslContextManager(), factory_context.statsScope())) {}
+
+ProtobufTypes::MessagePtr QuicClientTransportSocketConfigFactory::createEmptyConfigProto() {
+  return std::make_unique<envoy::extensions::transport_sockets::quic::v3::QuicUpstreamTransport>();
+}
+
+std::shared_ptr<quic::QuicCryptoClientConfig> QuicClientTransportSocketFactory::getCryptoConfig() {
+  Envoy::Ssl::ClientContextSharedPtr context = sslCtx();
+  // If the secrets haven't been loaded, there is no crypto config.
+  if (context == nullptr) {
+    ENVOY_LOG(warn, "SDS hasn't finished updating Ssl context config yet.");
+    stats_.upstream_context_secrets_not_ready_.inc();
+    return nullptr;
+  }
+
+  if (client_context_ != context) {
+    // If the context has been updated, update the crypto config.
+    client_context_ = context;
+    crypto_config_ = std::make_shared<quic::QuicCryptoClientConfig>(
+        std::make_unique<Quic::EnvoyQuicProofVerifier>(std::move(context)),
+        std::make_unique<quic::QuicClientSessionCache>());
+  }
+  // Return the latest crypto config.
+  return crypto_config_;
+}
+
+REGISTER_FACTORY(QuicClientTransportSocketConfigFactory,
+                 Server::Configuration::UpstreamTransportSocketConfigFactory);
+
+} // namespace Quic
+} // namespace Envoy
